integrate mersenne twister, importing javascript files, that dont export properly simply doesnt work. Either it works for webpack, or it works for the other runners/traditional require, but no way to make it work for both. Solution: integrate js as %x{}, works for everybody.
